refactor(trips): extract shared permission table widgets
Reuse a single header and row widget across trip general, participants, and expenses permission settings to remove duplicated layout code while preserving behavior.

Made-with: Cursor
