+++ lib/features/trips/presentation/trip_expenses_permissions_page.dart
@@ -5,10 +5,8 @@ import 'package:go_router/go_router.dart';
 import 'package:planerz/features/trips/data/trip_permission_helpers.dart';
 import 'package:planerz/features/trips/data/trip_permissions.dart';
 import 'package:planerz/features/trips/data/trips_repository.dart';
-import 'package:planerz/features/trips/presentation/widgets/permission_min_role_selector.dart';
+import 'package:planerz/features/trips/presentation/widgets/trip_permission_table_widgets.dart';
 import 'package:planerz/l10n/app_localizations.dart';
-
-const double _permissionsColumnSpacing = 12;
 
 class TripExpensesPermissionsPage extends ConsumerStatefulWidget {
   const TripExpensesPermissionsPage({
@@ -160,12 +158,12 @@ class _TripExpensesPermissionsPageState
                             ),
                       ),
                       const SizedBox(height: 12),
-                      _PermissionsColumnsHeader(
+                      TripPermissionsColumnsHeader(
                         actionLabel: l10n.tripPermissionsColumnAction,
                         minRoleLabel: l10n.tripPermissionsColumnMinRole,
                       ),
                       const SizedBox(height: 4),
-                      _ExpensesPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionExpensesCreatePost,
                         minRole: trip.expensesPermissions.createExpensePostMinRole,
                         icon: Icons.create_new_folder_outlined,
@@ -178,7 +176,7 @@ class _TripExpensesPermissionsPageState
                           minRole: role,
                         ),
                       ),
-                      _ExpensesPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionExpensesEditPost,
                         minRole: trip.expensesPermissions.editExpensePostMinRole,
                         icon: Icons.edit_outlined,
@@ -191,7 +189,7 @@ class _TripExpensesPermissionsPageState
                           minRole: role,
                         ),
                       ),
-                      _ExpensesPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionExpensesDeletePost,
                         minRole: trip.expensesPermissions.deleteExpensePostMinRole,
                         icon: Icons.delete_outline,
@@ -205,7 +203,7 @@ class _TripExpensesPermissionsPageState
                         ),
                       ),
                       const SizedBox(height: 20),
-                      _ExpensesPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionExpensesCreateExpense,
                         minRole: trip.expensesPermissions.createExpenseMinRole,
                         icon: Icons.add_circle_outline,
@@ -218,7 +216,7 @@ class _TripExpensesPermissionsPageState
                           minRole: role,
                         ),
                       ),
-                      _ExpensesPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionExpensesEditExpense,
                         minRole: trip.expensesPermissions.editExpenseMinRole,
                         icon: Icons.edit_note_outlined,
@@ -231,7 +229,7 @@ class _TripExpensesPermissionsPageState
                           minRole: role,
                         ),
                       ),
-                      _ExpensesPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionExpensesDeleteExpense,
                         minRole: trip.expensesPermissions.deleteExpenseMinRole,
                         icon: Icons.remove_circle_outline,
@@ -291,98 +289,3 @@ class _TripExpensesPermissionsPageState
   }
 }
 
-class _ExpensesPermissionItem extends StatelessWidget {
-  const _ExpensesPermissionItem({
-    required this.title,
-    required this.minRole,
-    required this.icon,
-    required this.busy,
-    required this.enabled,
-    required this.onChanged,
-  });
-
-  final String title;
-  final TripPermissionRole minRole;
-  final IconData icon;
-  final bool busy;
-  final bool enabled;
-  final ValueChanged<TripPermissionRole> onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 2),
-      child: Row(
-        children: [
-          Expanded(
-            flex: 6,
-            child: Row(
-              children: [
-                Icon(icon),
-                const SizedBox(width: 12),
-                Expanded(child: Text(title)),
-              ],
-            ),
-          ),
-          const SizedBox(width: _permissionsColumnSpacing),
-          Expanded(
-            flex: 4,
-            child: PermissionMinRoleSelector(
-              value: minRole,
-              busy: busy,
-              enabled: enabled,
-              onChanged: onChanged,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _PermissionsColumnsHeader extends StatelessWidget {
-  const _PermissionsColumnsHeader({
-    required this.actionLabel,
-    required this.minRoleLabel,
-  });
-
-  final String actionLabel;
-  final String minRoleLabel;
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-    final textStyle = Theme.of(context).textTheme.titleSmall?.copyWith(
-          color: colorScheme.onSecondaryContainer,
-          fontWeight: FontWeight.w700,
-        );
-
-    Widget buildCartouche(String label) {
-      return Container(
-        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
-        decoration: BoxDecoration(
-          color: colorScheme.secondaryContainer,
-          borderRadius: BorderRadius.circular(10),
-        ),
-        child: Text(label, style: textStyle),
-      );
-    }
-
-    return IntrinsicHeight(
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          Expanded(
-            flex: 6,
-            child: buildCartouche(actionLabel),
-          ),
-          const SizedBox(width: _permissionsColumnSpacing),
-          Expanded(
-            flex: 4,
-            child: buildCartouche(minRoleLabel),
-          ),
-        ],
-      ),
-    );
-  }
-}
