+++ lib/features/trips/presentation/trip_participants_permissions_page.dart
@@ -5,10 +5,8 @@ import 'package:go_router/go_router.dart';
 import 'package:planerz/features/trips/data/trip_permission_helpers.dart';
 import 'package:planerz/features/trips/data/trip_permissions.dart';
 import 'package:planerz/features/trips/data/trips_repository.dart';
-import 'package:planerz/features/trips/presentation/widgets/permission_min_role_selector.dart';
+import 'package:planerz/features/trips/presentation/widgets/trip_permission_table_widgets.dart';
 import 'package:planerz/l10n/app_localizations.dart';
-
-const double _permissionsColumnSpacing = 12;
 
 class TripParticipantsPermissionsPage extends ConsumerStatefulWidget {
   const TripParticipantsPermissionsPage({
@@ -153,12 +151,12 @@ class _TripParticipantsPermissionsPageState
                             ),
                       ),
                       const SizedBox(height: 12),
-                      _PermissionsColumnsHeader(
+                      TripPermissionsColumnsHeader(
                         actionLabel: l10n.tripPermissionsColumnAction,
                         minRoleLabel: l10n.tripPermissionsColumnMinRole,
                       ),
                       const SizedBox(height: 4),
-                      _ParticipantsPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionParticipantsCreate,
                         minRole:
                             trip.participantsPermissions.createParticipantMinRole,
@@ -172,7 +170,7 @@ class _TripParticipantsPermissionsPageState
                           minRole: role,
                         ),
                       ),
-                      _ParticipantsPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionParticipantsEditPlaceholder,
                         minRole: trip
                             .participantsPermissions
@@ -188,7 +186,7 @@ class _TripParticipantsPermissionsPageState
                           minRole: role,
                         ),
                       ),
-                      _ParticipantsPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionParticipantsDeletePlaceholder,
                         minRole: trip
                             .participantsPermissions
@@ -205,7 +203,7 @@ class _TripParticipantsPermissionsPageState
                           minRole: role,
                         ),
                       ),
-                      _ParticipantsPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionParticipantsDeleteRegistered,
                         minRole: trip
                             .participantsPermissions
@@ -222,7 +220,7 @@ class _TripParticipantsPermissionsPageState
                           minRole: role,
                         ),
                       ),
-                      _ParticipantsPermissionItem(
+                      TripPermissionItemRow(
                         title: l10n.tripPermissionParticipantsToggleAdmin,
                         minRole:
                             trip.participantsPermissions.toggleAdminRoleMinRole,
@@ -283,98 +281,3 @@ class _TripParticipantsPermissionsPageState
   }
 }
 
-class _ParticipantsPermissionItem extends StatelessWidget {
-  const _ParticipantsPermissionItem({
-    required this.title,
-    required this.minRole,
-    required this.icon,
-    required this.busy,
-    required this.enabled,
-    required this.onChanged,
-  });
-
-  final String title;
-  final TripPermissionRole minRole;
-  final IconData icon;
-  final bool busy;
-  final bool enabled;
-  final ValueChanged<TripPermissionRole> onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 2),
-      child: Row(
-        children: [
-          Expanded(
-            flex: 6,
-            child: Row(
-              children: [
-                Icon(icon),
-                const SizedBox(width: 12),
-                Expanded(child: Text(title)),
-              ],
-            ),
-          ),
-          const SizedBox(width: _permissionsColumnSpacing),
-          Expanded(
-            flex: 4,
-            child: PermissionMinRoleSelector(
-              value: minRole,
-              busy: busy,
-              enabled: enabled,
-              onChanged: onChanged,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _PermissionsColumnsHeader extends StatelessWidget {
-  const _PermissionsColumnsHeader({
-    required this.actionLabel,
-    required this.minRoleLabel,
-  });
-
-  final String actionLabel;
-  final String minRoleLabel;
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-    final textStyle = Theme.of(context).textTheme.titleSmall?.copyWith(
-          color: colorScheme.onSecondaryContainer,
-          fontWeight: FontWeight.w700,
-        );
-
-    Widget buildCartouche(String label) {
-      return Container(
-        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
-        decoration: BoxDecoration(
-          color: colorScheme.secondaryContainer,
-          borderRadius: BorderRadius.circular(10),
-        ),
-        child: Text(label, style: textStyle),
-      );
-    }
-
-    return IntrinsicHeight(
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          Expanded(
-            flex: 6,
-            child: buildCartouche(actionLabel),
-          ),
-          const SizedBox(width: _permissionsColumnSpacing),
-          Expanded(
-            flex: 4,
-            child: buildCartouche(minRoleLabel),
-          ),
-        ],
-      ),
-    );
-  }
-}
